pyproject.toml
[project]
name = "lab-extractor"
version = "0.1.0"
description = "Agentic lab report structured data extraction using MedGemma 1.5 4B"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.50.0",
    "torch>=2.0.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "accelerate>=0.26.0",
    "pdf2image>=1.16.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
